Datatypes

Description

Documentation

data User

Versão resumida de usuário do sistema. Mantém dados essenciais, mas omite dados sensíveis que não serão necessários.

Constructors

User

Fields

nameUser :: String

O nome do Usuário.

emailUser :: String

O e-mail do usuário.

isAdminUser :: Bool

Indicação se o usuário é ou não um administrador.

data UserFull

Versão completa do usuário do sistema, a ser usada quando se precisa de todas as informações do usuário.

Constructors

UserFull

Fields

name :: String

O nome do usuário.

email :: String

O e-mail do usuário.

password :: Text

A senha (criptografada) do usuário.

registrationDate :: String

A data em que o usuário se registrou.

isAdmin :: Bool

Indicação se o usuário é administrador.

type ErrorLog = String

Define o tipo ErrorLog como uma String.

data Screen

As diferentes telas do sistema.

Constructors

ExitScreen
A tela de saída do sistema.

FirstScreen
A tela inicial, quando nenhum administrador foi cadastrado ainda.

StartScreen
A tela inicial do sistema quando já existem usuários cadastrados.

LoginScreen
A tela de Login.

LoggedScreen
A tela após um usuário logar-se no sistema.

AdminScreen
A tela exibida após um administrador fazer login no sistema.

InsertRoomScreen
A tela de cadastrar uma nova sala no sistema

SignUpScreen
A tela de cadastro do usuário.

RemoveUserScreen
A tela de remoção de um usuário.

ViewUserScreen
A tela de visualização de salas específica de um usuário.

ViewScreen
A tela com opções de visualização de salas.

ViewRoomScreen
A tela de visualização de uma sala específica.

ViewFilterScreen
A tela de busca de salas por filtros.

ReportRoomScreen
A tela de exibição de relatório de ocupação para uma sala específica.

ReportDayScreen
A tela de exibição de relatório de ocupação para todas as salas em um determinado dia.

CreateReservationScreen
A tela para se fazer uma nova reserva em uma sala.

EditReservationScreen
A tela para se editar uma reserva previamente feita.

RemoveReservationScreen
A tela para se remover uma reserva previamente feita.

data Room

Uma sala do sistema SIGES.

Constructors

Room

Fields

code :: String

O código da sala.

schedule :: [Reservation]

A agenda onde são armazenadas todas as reservas ativas para esta sala.

resources :: [Resource]

A lista de recursos oferecidos pela sala.

capacity :: Int

A quantidade de pessoas suportada pela sala.

localization :: String

A Localização da Sala.

category :: RoomCategory

A categoria da sala.

data Reservation

As reservas de salas do sistema.

Constructors

Reservation

Fields

requester :: String

O responsável pela reserva.

description :: String

A descrição do evento da reserva.

startTime :: LocalTime

O tempo de início do evento.

finishTime :: LocalTime

O tempo de finalização do evento.

data RoomCategory

As categorias de sala oferecidas pelo sistema.

Constructors

Laboratory
Sala que funciona como um laboratório.

Auditorium
Sala que funciona como um auditório.

Classroom
Sala que funciona como uma sala de aula.

Office
Sala que funciona como um escritório.

Warehouse
Sala que funciona como um depósito.

data ResourceKind

Os tipos de recursos existentes no sistema.

Constructors

Projector
Projetor ou Datashow.

Microscope
Microscópio.

Desk
Escrivaninha.

Computer
Computador.

Board
Quadro negro ou branco.

AirConditioner
Ar condicionado.

data Resource

Os recursos oferecidos pelas salas do sistema.

Constructors

Resource

Fields

resourceKind :: ResourceKind

O tipo do recurso.

resourceQuantity :: Int

A quantidade do recurso que a sala oferece.