Handlers.DataHandler

Description

Documentation

userTmpJSON :: FilePath

Esta função tem o valor igual ao caminho para o arquivo json com os dados temporários de usuário.

userJSON :: FilePath

Esta função tem o valor igual ao caminho para o arquivo json com os dados de usuário.

fileExists :: FilePath -> IO Bool

Dado um FilePath, esta função verifica se ele se refere a um arquivo existente.

noUsersYet :: IO Bool

Esta função verifica os usuários cadastrados no sistema e retornará um valor booleano true caso o sistema não possua nenhum usuário cadastrado ainda, e false em caso contrário.

saveUser :: UserFull -> IO Bool

Esta função considera um UserFull, e caso o seu e-mail já não esteja cadastrado no sistema,. ela o incluirá no sistema, retornando um valor booleano indicando se foi possível incluir o novo usuário.

deleteUser :: String -> IO Bool

Esta função considera uma String, e caso ela corresponda ao e-mail de um dos usuários, a função eliminará o usuário equivalente do sistema, retornando um valor booleano indicando se a remoção pôde ser feita.

getUser :: String -> IO (Maybe UserFull)

Esta função considerará uma String, e retornará o usuário cadastrado no sistema com o e-mail igual a esta String, caso exista.

signUser :: String -> IO ()

Esta função considerará uma String, e verificará o sistema para buscar o usuário cadastrado no sistema com o e-mail igual à String. O UserFull será então incluído, como Uswer (Que contém menos informações) nos dados temporários do sistema. Estes dados poderão então ser acessados em tempo de execução do sistema.

hasLoggedUser :: IO Bool

Esta função decidirá se algum usuário está atualmente logado no sistema.

getLoggedUser :: IO User

Esta função retornará o Usuário atualmente logado no sistema.

signOutUser :: IO ()

Esta função deslogará um usuário do sistema, destruindo as informações salvas temporariamente.

roomsJSON :: FilePath

Esta função tem o valor igual ao caminho para o arquivo json com os dados das salas.

fetchRooms :: IO [Room]

Esta função retorna uma lista contendo todas as salas cadastradas no sistema.

noRoomsYet :: IO Bool

Esta função verifica as salas cadastradas no sistema e retornará um valor booleano true caso o sistema não possua nenhuma sala cadastrada ainda, e false em caso contrário.

saveRoom :: Room -> IO Bool

Esta função considera uma sala, e caso ela ainda não exista no sistema, será incluída e um valor booleano true será retornado. Caso contrário, um valor booleano false será retornado.

updateRoom :: String -> Room -> IO Bool

Esta função usa uma String contendo um código de sala e uma sala nova para substituir a sala com o código dado pela versão nova da mesma sala.

updateAllRooms :: (Room -> Room) -> IO Bool

Esta função recebe uma função e atualiza todas as salas do sistema aplicando sobre elas a função determinada.

deleteRoom :: String -> IO Bool

Esta função considera uma String, e caso ela corresponda ao código de uma das salas, a função eliminará a sala equivalente do sistema, retornando um valor booleano indicando se a remoção pôde ser feita.

getRoom :: String -> IO (Maybe Room)

Esta função considerará uma String, e retornará a sala cadastrada no sistema com o código igual a esta String, caso exista.

Orphan instances

ToJSON Resource

Instance details

Methods

toJSON :: Resource -> Value

toEncoding :: Resource -> Encoding

toJSONList :: [Resource] -> Value

toEncodingList :: [Resource] -> Encoding

ToJSON ResourceKind

Instance details

Methods

toJSON :: ResourceKind -> Value

toEncoding :: ResourceKind -> Encoding

toJSONList :: [ResourceKind] -> Value

toEncodingList :: [ResourceKind] -> Encoding

ToJSON RoomCategory

Instance details

Methods

toJSON :: RoomCategory -> Value

toEncoding :: RoomCategory -> Encoding

toJSONList :: [RoomCategory] -> Value

toEncodingList :: [RoomCategory] -> Encoding

ToJSON Reservation

Instance details

Methods

toJSON :: Reservation -> Value

toEncoding :: Reservation -> Encoding

toJSONList :: [Reservation] -> Value

toEncodingList :: [Reservation] -> Encoding

ToJSON Room

Instance details

Methods

toJSON :: Room -> Value

toEncoding :: Room -> Encoding

toJSONList :: [Room] -> Value

toEncodingList :: [Room] -> Encoding

ToJSON UserFull

Instance details

Methods

toJSON :: UserFull -> Value

toEncoding :: UserFull -> Encoding

toJSONList :: [UserFull] -> Value

toEncodingList :: [UserFull] -> Encoding

ToJSON User

Instance details

Methods

toJSON :: User -> Value

toEncoding :: User -> Encoding

toJSONList :: [User] -> Value

toEncodingList :: [User] -> Encoding

FromJSON Resource

Instance details

Methods

parseJSON :: Value -> Parser Resource

parseJSONList :: Value -> Parser [Resource]

FromJSON ResourceKind

Instance details

Methods

parseJSON :: Value -> Parser ResourceKind

parseJSONList :: Value -> Parser [ResourceKind]

FromJSON RoomCategory

Instance details

Methods

parseJSON :: Value -> Parser RoomCategory

parseJSONList :: Value -> Parser [RoomCategory]

FromJSON Reservation

Instance details

Methods

parseJSON :: Value -> Parser Reservation

parseJSONList :: Value -> Parser [Reservation]

FromJSON Room

Instance details

Methods

parseJSON :: Value -> Parser Room

parseJSONList :: Value -> Parser [Room]

FromJSON UserFull

Instance details

Methods

parseJSON :: Value -> Parser UserFull

parseJSONList :: Value -> Parser [UserFull]

FromJSON User

Instance details

Methods

parseJSON :: Value -> Parser User

parseJSONList :: Value -> Parser [User]