Handlers.ErrorHandler

Description

Documentation

validScreen :: Screen -> String -> IO (Either ErrorLog Screen)

Dada uma determinada tela do sistema e uma resposta, esta função verifica se a resposta foi adequada àquela tela, retornando, em caso positivo, a tela equivalente à resposta, ou em caso negativo, o ErrorLog com a mensagem identificando qualerro existente na resposta.

singleCharIsR :: String -> Bool

Dada uma String, esta função verificará se esta String é composta unicamente pelo caractere R, respondendo com um valor booleano.

checkNewEmail :: String -> IO (Either ErrorLog String)

Esta função considerará um e-mail, em String, e decidirá se ele é válido e livre ou se é inválido ou se já está em uso, retornando uma mensagem de erro caso o e-mail não seja válido e livre.

checkEmail :: String -> IO (Either ErrorLog String)

Esta função considerará um e-mail, em String, e decidirá se ele é válido e está cadastrado. Caso não seja válido ou não esteja cadastrado, retornará uma mensagem de erro. Caso a função receba como parâmetro uma String composta apenas da letra R, responderá com a String Retornar.

checkNewPass :: String -> IO (Either ErrorLog String)

Esta função considerará uma String determinada e decidirá se esta String pode ser usada como senha neste sistema. Em caso negativo, retornará uma mensagem de erro.

checkPass :: String -> String -> IO (Either ErrorLog String)

Esta função considera um e-mail e uma senha, e decide se a senha está correta (caso em que retorna a própria senha) ou não (Caso em que retorna a mensagem de erro correspondente.)

checkName :: String -> IO (Either ErrorLog String)

Esta função considera uma String e decide se ela pode ser considerada como um nome do usuário (apenas letras e espaços são permitidos). Caso a String contenha um nome válido, o próprio nome será retornado. Caso contrário, uma mensagem de erro será retornada.

yesOrNo :: String -> IO (Either ErrorLog Bool)

Esta função considera uma String e decide se ela representa corretamente uma resposta de Sim ou Não (S ou N). Caso tenha mais de um caractere ou não seja S ou N, retornará uma mensagem de erro. Caso seja válida, a função retornará um Booleano. (True para S e False para N).

checkNewRoomCode :: String -> IO (Either ErrorLog String)

Esta função considera uma Sting e decide se ela equivale ao código de uma das salas cadastradas no sistema (Caso em que retornará uma mensagem de erro). Caso contrário, a função retornará em que retornará o próprio código.

checkRoomCode :: String -> IO (Either ErrorLog String)

Esta função considera uma Sting e decide se ela equivale ao código de uma das salas cadastradas no sistema (Caso em que retornará o próprio código). Caso contrário, a função retornará uma mensagem de erro.

checkDay :: String -> IO (Either ErrorLog [Int])

Esta função considera uma String, e decidirá se ela equivale a uma data válida. Em caso positivo, retornará o mesmo na forma de uma lista de inteiros. Caso contrário, retornará uma mensagem de erro equivalente ao problema encontrado.

checkTime :: String -> IO (Either ErrorLog [Int])

Esta função considera uma String e decide se ela equivale a um horário. Em caso positivo, retornará o horário em forma de Lista de inteiros. Caso contrário, retornará a mensagem de erro equivalente ao problema encontrado.

checkDescription :: String -> IO (Either ErrorLog String)

Esta função considera uma String e decide se ela pode ser usada como descrição para uma reserva de sala. Em caso positivo, retornará a mesma String. Caso contrário, retornará uma mensagem de erro.

checkCategory :: String -> IO (Either ErrorLog RoomCategory)

Esta função considera uma String e decide se ela representa uma das categorias no sistema. Caso represente, a função retornará a Categoria desejada, e em caso contrário, retornará uma mensagem de erro.

checkNumber :: String -> IO (Either ErrorLog Int)

Esta função considera uma String e decide se ela contém um valor inteiro entre zero e trezentos, caso no qual retornará o inteiro. Em caso contrário, uma mensagem de erro será retornada.

checkResource :: String -> IO (Either ErrorLog ResourceKind)

Esta função considera uma String e decide se ela representa um dos tipos de recursos oferecidos pelas salas do sistema. Caso afirmativo, a função retornará o ResourceKind solicitado. Caso contrário, uma mensagem de erro será retornada.

checkFilter :: String -> IO (Either ErrorLog Int)

Esta função considera uma String e decide se ela equivale a uma das opções (1, 2, 3, ou 4), caso em que retorna a opção como um inteiro. Em caso contrário, retornará uma mensagem de erro.