Handlers.RoomsHandler

Description

Documentation

createRoom :: String -> [Resource] -> RoomCategory -> Int -> String -> IO Bool

Função que, dadas todas as informações de uma sala, a cria e a persiste no banco de dados, retornando o booleano que representa o sucesso da operação

makeTime :: (Integer, Int, Int, Int, Int) -> LocalTime

Função auxiliar que, dada uma tupla com um Integer para o ano e um Int para mÊs, dia, hora e minuto, respectivamente, cria um LocalTime.

isFree :: Room -> LocalTime -> Bool

Esta função recebe uma sala e um horário, e verifica se esta sala estará livre neste horário, respondendo com um valor booleano.

makeReservation :: String -> String -> String -> (Integer, Int, Int, Int, Int) -> (Integer, Int, Int, Int, Int) -> IO Bool

Esta função criará uma reserva em uma das salas, a partir do código da mesma, do nome do responsável pela reserva, e das tuplas especificando data e horário de finalização do evento.

deleteReservation :: String -> String -> (Integer, Int, Int, Int, Int) -> IO Bool

Esta função deletará uma função identificada pelo código da sala e nome do responsável, e pela tupla do horário de início do evento.

findReservation :: String -> (Integer, Int, Int, Int, Int) -> String -> IO Reservation

Dada uma sala, identificada pelo seu código, um horário, representado em forma de tupla e um nome de usuário, esta função retornará a reserva com os dados equivalentes.

findReservationEasy :: String -> (Integer, Int, Int, Int, Int) -> IO Reservation

Dada uma sala, identificada pelo seu código, e um horário em forma de tupla, esta função retornará a reserva da sala cujo evento se inicie no horário dado.

editReservation :: String -> String -> (Integer, Int, Int, Int, Int) -> (Integer, Int, Int, Int, Int) -> (Integer, Int, Int, Int, Int) -> IO Bool

Esta função alterará o horário de uma reserva (identificada pelo código de sua sala, pelo seu responsável e horário de início). Caso não seja possível fazer a alteração, nada será feito. A função retornará um valor booleano indicando se foi possível fazer a operação.

cleanReservations :: LocalTime -> Room -> Room

Dado um LocalTime para comparação, esta função irá retirar de uma determinada sala as reservas finalizadas antes deste horário.

cleanAllReservations :: IO Bool

Esta função removerá de todas as salas as reservas cujo horário de final do evento já passou.

createReportForTheRoom :: (Integer, Int, Int) -> Room -> String

Dado um dia representado em uma tupla e um código de sala, esta função criará um relatório em texto com todas as reservas que esta sala tem para o dia especificado.

createReportForTheDay :: (Integer, Int, Int) -> IO String

Dado um dia representado em uma tupla, esta função criará um relatório em texto para todas as reservas de todas as salas para o dia especificado.

searchRoomsCategory :: RoomCategory -> IO [Room]

Com uma categoria especificada, esta função verificará o sistema e retornará a lista contendo todas as salas desta categoria.

searchRoomsCapacity :: Int -> IO [Room]

Com uma capacidade especificada, esta função verificará o sistema e retornará a lista contendo todas as salas com esta capacidade ou mais.

searchRoomsTime :: (Integer, Int, Int, Int, Int) -> (Integer, Int, Int, Int, Int) -> IO [Room]

Com um horário de início e fim especificado em forma de tupla, esta função verificará o sistema e retornará a lista contendo todas as salas que estejam livres neste horário.

resourceIsEnough :: Room -> Resource -> Bool

Esta função recebe uma sala e um Resource e verifica se esta sala tem o recurso e se o tem na quantidade especificada ou superior, retornando a resposta em forma de valor booleano.

containsResources :: [Resource] -> Room -> Bool

Esta função verificará se uma sala contém todos os recursos em uma lista, e se os tem na quantidade especificada ou superior.

searchRoomsResources :: [Resource] -> IO [Room]

Com uma lista de recursos especificada, esta função verificará o sistema e retornará a lista contendo todas as salas que supram esta demanda.

wasReservedBy :: String -> Room -> Bool

Dado um nome de usuário e uma sala, esta função verifica se a sala foi reservada por este usuário para algum horário e retorna a resposta em forma de valor booleano.

searchRoomsRequester :: String -> IO [Room]

Com um nome de usuário especificado, esta função verificará o sistema e retornará a lista contendo todas as salas que foram reservadas para algum horário por aquele usuário.

printCategories :: IO ()

Esta função produzirá um texto contendo a lista de todas as categorias de sala suportadas pelo sistema.

printResources :: IO ()

Esta função produzirá um texto listando todos os recursos oferecidos pelas salas do sistema.

Orphan instances

Ord Reservation

Instance details

Methods

compare :: Reservation -> Reservation -> Ordering

(<) :: Reservation -> Reservation -> Bool

(<=) :: Reservation -> Reservation -> Bool

(>) :: Reservation -> Reservation -> Bool

(>=) :: Reservation -> Reservation -> Bool

max :: Reservation -> Reservation -> Reservation

min :: Reservation -> Reservation -> Reservation

Show Resource

Instance details

Methods

showsPrec :: Int -> Resource -> ShowS

show :: Resource -> String

showList :: [Resource] -> ShowS

Show Reservation

Instance details

Methods

showsPrec :: Int -> Reservation -> ShowS

show :: Reservation -> String

showList :: [Reservation] -> ShowS

Show Room

Instance details

Methods

showsPrec :: Int -> Room -> ShowS

show :: Room -> String

showList :: [Room] -> ShowS