Handlers.UserHandler

Description

Documentation

userExists :: String -> IO Bool

Esta função considera uma String e decide se um usuário com o e-mail igual a esta String existe no sistema.

makePass :: String -> Password

Dada uma String, esta função a converterá em um Password.

correctPassword :: String -> String -> IO Bool

Esta função considerará uma String contendo um e-mail e avaliará outra String, decidindo se ela equivale à senha do usuário com o e-mail fornecido. A resposta é então retornada em forma de valor booleano.

retrieveUser :: String -> IO User

Dada uma String contendo um e-mail, esta função procurará no sistema um Userfull com este e-mail, e retornará a sua versão com as informações essenciais para o funcionamento do sistema, omitindo dados sensíveis: um User.

registerNewUser :: String -> String -> String -> Bool -> IO ()

Dadas três Strings contendo respectivamente o e-mail, a senha e o nome do usuário, e um valor booleano indicando se trata-se de um administrador, esta função cria um novo usuário e o armazena no sistema.

removeUser :: User -> IO ()

Esta função recebe um usuário e o remove do sistema.

Orphan instances

Show User

Instance details

Methods

showsPrec :: Int -> User -> ShowS

show :: User -> String

showList :: [User] -> ShowS