TUI.ScreenListeners

Description

Documentation

class Action a where

Typeclass que estabelece as operações de ações a ser executadas em cada tela do sistema.

Methods

useContent

Arguments

:: a

-> IO Screen
A implementação de como a tela será impressa na tela e como lerá as informações dadas pelo usuário.

userInteraction :: Screen -> IO Screen

Esta é a função responsável por encontrar e controlar quais telas serão exibidas no sistema SIGES.

waitInput :: IO String

getAnswer :: IO String

Esta função exibirá um prompt e captará a resposta digitada pelo usuário.

getPassword :: IO String

Função responsável por captar a senha do usuário da entrada e garantir que ela seja exibida apenas com asteriscos, como forma de melhorar a segurança.

getInputData :: IO String -> (String -> IO (Either ErrorLog a)) -> IO a

Dada uma Mônada de String qualquer e uma função de verificação, esta função aplicará a função sobre a String. Caso a String passe na verificação fornecida, será retornada. Caso contrário, uma mensagem de erro será exibida e a função será repetida com uma nova entrada.

getRoomsFilter :: [Room] -> IO [Room]

Esta filtrará salas pelos filtros escolhidos pelo usuário. MAis de um filtro podem ser aplicados simultaneamente, neste caso, a função perguntará um a um quais filtros devem ser adicionados.

getResources :: IO [Resource]

Função auxiliar à função de filtro de salas. Esta função lerá da entrada recursos que o usuário deseja buscar e suas respectivas quantidades, e então produzirá uma lista de recursos com estas informações e a retornará.